backend/requirements.txt
Flask==2.2.2
pymongo==4.3.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
Flask-JWT-Extended==4.4.4
face_recognition==1.3.0
numpy==1.24.2
Pillow==9.4.0
bcrypt==4.0.1
openpyxl==3.1.2